backend/requirements.txt
django>=5.0
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
openai
stripe
google-generativeai